+++ src/strings_utils_base.hpp
@@ -1,0 +1,13 @@
+// Copyrighted (C) 2017-2017 Antinet.org team, see file LICENCE-by-Antinet.txt
+
+#pragma once
+
+enum t_debug_style {
+	e_debug_style_object=0,
+	e_debug_style_short_devel=1,
+	e_debug_style_crypto_devel=2,
+	e_debug_style_big=2,
+
+	e_debug_style_buf=100,
+};
+
