+++ src/test/Makefile
@@ -2,5 +2,5 @@
 .PHONY: run
 
 run:
-	( cd ../../build && make && ./test-debug.elf --gtest_filter=stdplus_tab.simple )
+	( cd ../../build && make && ./test-debug.elf --gtest_filter=stdplus_tab.* )
